Fixed Gene structure creation
Fixed up the creation of structures by the genes in a genome. Also setup examples of the genome in the cell. I would like to setup a "Starting Genome" that the user can edit before the start of the simulation. We can have it show an example of the cell given the genome provided